fix: fix uris param name, use DateTimeFormatter on controller aram

+++ stats-service/stats-dto/src/main/java/stats/ViewStats.java
@@ -12,5 +12,5 @@ import lombok.NoArgsConstructor;
 public class ViewStats {
     private String app;
     private String uri;
-    private Long hits;
+    private int hits;
 }
